pyproject.toml
[tool.poetry]
name = "telegram-channel-bot"
version = "0.1.0"
description = "Telegram bot for VIP and free channel management"
authors = ["Your Name <your @email.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.4.1"
python-dotenv = "^1.0.1"
SQLAlchemy = "^2.0.30"
alembic = "^1.13.1"
asyncpg = "^0.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"